# core-lending-asset-finance-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /api/v1/asset-finance-agreements/{agreementId}/end-options/{optionId}:
    get:
      tags:
      - EndOption
      summary: Get an end option by ID
      description: Retrieve a specific lease-end purchase option by its unique identifier
      operationId: getById_1
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: optionId
        in: path
        description: Unique identifier of the end option
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: End option found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndOptionDTO"
        "404":
          description: End option or agreement not found
        "500":
          description: Internal server error
    put:
      tags:
      - EndOption
      summary: Update an existing end option
      description: Update all fields of an existing lease-end purchase option
      operationId: update_1
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: optionId
        in: path
        description: Unique identifier of the end option to update
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndOptionDTO"
        required: true
      responses:
        "200":
          description: End option updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndOptionDTO"
        "400":
          description: Invalid input data
        "404":
          description: End option or agreement not found
        "500":
          description: Internal server error
    delete:
      tags:
      - EndOption
      summary: Delete an end option
      description: Permanently delete a lease-end purchase option
      operationId: delete_1
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: optionId
        in: path
        description: Unique identifier of the end option to delete
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "204":
          description: End option deleted successfully
        "404":
          description: End option or agreement not found
        "500":
          description: Internal server error
...
  /api/v1/asset-finance-agreements/{agreementId}/end-options:
    get:
      tags:
      - EndOption
      summary: List/Search end options for an agreement
      description: Retrieve a paginated list of lease-end purchase options for a specific
        agreement
      operationId: findAll_1
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: filterRequest
        in: query
        description: Filter criteria for searching end options
        required: true
        schema:
          $ref: "#/components/schemas/FilterRequestEndOptionDTO"
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved list of end options
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginationResponseEndOptionDTO"
        "400":
          description: Invalid filter request
        "404":
          description: Agreement not found
        "500":
          description: Internal server error
    post:
      tags:
      - EndOption
      summary: Create a new end option for an agreement
      description: Create a new lease-end purchase option for an agreement
      operationId: create_1
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndOptionDTO"
        required: true
      responses:
        "200":
          description: End option created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndOptionDTO"
        "400":
          description: Invalid input data
        "404":
          description: Agreement not found
        "500":
          description: Internal server error
...
components:
  schemas:
...
    EndOptionDTO:
      required:
      - assetFinanceAgreementId
      - optionExerciseDate
      - optionPaidAmount
      type: object
      properties:
        endOptionId:
          type: string
          format: uuid
          readOnly: true
        assetFinanceAgreementId:
          type: string
          format: uuid
        optionExerciseDate:
          type: string
          format: date
        optionPaidAmount:
          type: number
        isExercised:
          type: boolean
        note:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: "The direction of sorting, either ASC or DESC."
          example: DESC
          default: DESC
      description: "Represents a pagination request for retrieving paginated results,\
        \ including page number, size, sort field, and direction."
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/RangeObject"
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    FilterRequestEndOptionDTO:
      required:
      - pagination
      type: object
      properties:
        filters:
          $ref: "#/components/schemas/EndOptionDTO"
        rangeFilters:
          $ref: "#/components/schemas/RangeFilter"
        pagination:
          $ref: "#/components/schemas/PaginationRequest"
        options:
          $ref: "#/components/schemas/FilterOptions"
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseEndOptionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: "#/components/schemas/EndOptionDTO"
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: "The current page number, typically zero-based."
          format: int32
      description: Represents a paginated response containing a list of items and
        pagination metadata.
...